+++ scripts/build.py
@@ -12,8 +12,7 @@ import sys
 CRATE_NAME = "pixelscript"
 LIB_CRATES = [CRATE_NAME]
 SOURCE = "pxsb"
-# TODO: Figure out how WASM will work since it needs to use other libs for it to work, gotta have to link something?
-VALID_EXTENSIONS = ["lib", "a", "so", "dylib"]
+VALID_EXTENSIONS = ["lib", "a", "so", "dylib", "wasm", "dll", "js"]
 full_lib_size = 0
 
 
@@ -32,6 +31,9 @@ def move(old):
     old = convert_path(old)
     ext = get_ext(old)
     file_name = old.split('.')[0].split('/')[-1]
+    # The wasm build produces pxs.js and pxs.wasm
+    if file_name == 'pxs' and ext in ['wasm', 'js']:
+        file_name = "pixelscript"
     shutil.copy(old, f"{SOURCE}/{file_name}.{ext}")
 
 
@@ -54,6 +56,8 @@ defaults = True
 debug = False
 run_clear = False
 use_zig = False
+two_wasm = False
+binary = False
 
 for arg in argv:
     if "target" in arg:
@@ -72,6 +76,10 @@ for arg in argv:
         run_clear = True
     elif arg == "zig":
         use_zig = True
+    elif arg == '2wasm':
+        two_wasm = True
+    elif arg == 'bin':
+        binary = True
     elif arg == "help":
         print("""PixelScript script/build.py usage
 Arguments:
@@ -82,13 +90,26 @@ Arguments:
 - clear; clear the cache
 - help; print this message
 - zig; use zig to build. This is best for cross platform. Requires `cargo-zigbuild`.
+- 2wasm; comple pixelscript library for use in WASM.
+- bin; compile the pxs binary.
 """)
         exit(0)
 
 build_mode = "release" if not debug else "debug"
 build_flag = "--release" if not debug else ""
 # Build in release mode
-cmd = ["cargo", "zigbuild" if use_zig else "build", build_flag]
+cmd = ["cargo", "zigbuild" if use_zig else "build", build_flag, "--lib" if not binary else "--bin pxs"]
+
+# override target when wasm is passed. 
+if two_wasm:
+    if not binary:
+        print("Can not compile WASM without `bin` enabled.")
+        exit(1)
+    if use_zig:
+        print("Can not compile WASM using zig.")
+        exit(1)
+    target = "--target=wasm32-unknown-emscripten"
+    rtarget = "wasm32-unknown-emscripten"
 # Grab target and features if passed
 if target:
     cmd += [target]
@@ -96,7 +117,6 @@ if not defaults:
     cmd += ["--no-default-features"]
 if len(features) > 0:
     cmd += ["--features", f'"{features}"']
-# cmd += features
 
 if run_clear:
     # Run cargo clean before building
@@ -132,5 +152,17 @@ for path in os.listdir(build_dir):
             print(full_path)
             # Search through contents
             collect_libs(full_path)
+
+# Replace pxs.wasm to pixelscript.wasm in pixelscript.js
+if two_wasm:
+    files = glob(f"{SOURCE}/*.js")
+    if len(files) > 0:
+        with open(files[0], "r+") as f:
+            contents = f.read()
+            contents = contents.replace("pxs.wasm", "pixelscript.wasm")
+            f.seek(0)
+            f.truncate(0)
+            f.write(contents)
+
 
 print(f"Full size of pixelscript: {full_lib_size // 1000000}mb")