+++ src/pxs_core/pxs_zip.rs
@@ -1,10 +1,10 @@
-use crate::shared::module::pxs_Module;
+// use crate::shared::module::pxs_Module;
 
-unsafe extern "C" {
-    /// cbindgen:ignore
-    fn pxs_corelib_zip_init();
-}
+// unsafe extern "C" {
+//     /// cbindgen:ignore
+//     fn pxs_corelib_zip_init();
+// }
 
-pub(crate) fn init(_module: *mut pxs_Module) {
-    unsafe { pxs_corelib_zip_init() };
-}
+// pub(crate) fn init(_module: *mut pxs_Module) {
+//     unsafe { pxs_corelib_zip_init() };
+// }